streamlit/requirements.txt
streamlit>=1.28.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
plotly>=5.0.0
pytz>=2021.1
streamlit-plotly-events>=0.0.6
streamlit-lightweight-charts-ntf>=0.8.0
pandas>=2.0.0
numpy>=1.24.0,<2.0.0
requests>=2.25.0